+++ codeship/resource_encrypted.go
@@ -9,7 +9,6 @@ import (
 	"io"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/diag"
-
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 )